% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function-template.R
\name{function_template}
\alias{function_template}
\title{Function Template}
\usage{
function_template(x = TRUE, chk = TRUE)
}
\arguments{
\item{x}{A flag specifying whether to return TRUE or FALSE.}

\item{chk}{A flag specifying whether to check inputs.}
}
\value{
A flag.
}
\description{
A function template.
}
\examples{
function_template()
function_template(FALSE)
}
