% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{pkgtemplate-package}
\alias{pkgtemplate}
\alias{pkgtemplate-package}
\title{pkgtemplate: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/audrey-b/pkgtemplate}
  \item Report bugs at \url{https://github.com/audrey-b/pkgtemplate/issues}
}

}
\author{
\strong{Maintainer}: Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca}

}
\keyword{internal}
