% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{simanalyse-package}
\alias{simanalyse}
\alias{simanalyse-package}
\title{simanalyse: Facilitate Model Comparisons and Simulation Studies}
\description{
Facilitates model comparisons and simulation studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/audrey-b/simanalyse}
  \item Report bugs at \url{https://github.com/audrey-b/simanalyse/issues}
}

}
\author{
\strong{Maintainer}: Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca}

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
}

Other contributors:
\itemize{
  \item Harmohit Singh Bindra \email{hsbindra@uwaterloo.ca} [contributor]
}

}
\keyword{internal}
