% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-analyse-bayesian.R
\name{simanalyse_analyse_bayesian}
\alias{simanalyse_analyse_bayesian}
\title{Analyse data for a simulation study}
\usage{
simanalyse_analyse_bayesian(datalist, code, code_add = NULL,
  code_values = "", package, monitor, inits, n.adapt, n.burnin, n.iter,
  thin, seed, path, exists, silent, ...)
}
\arguments{
\item{datalist}{A list of nlist objects containing the data. Alternatively, path can be used to specify to read the data from files.}

\item{code}{A string of code to analyse the data. JAGS code must not be in a data or model block.}

\item{code_add}{A string of code to append at the end of \code{code} before analysing the data. This is useful for specifying priors seperately from the likelihood in JAGS.}

\item{code_values}{A character vector to pass to sprintf when applied to the appended code from \code{code} and \code{code_add}}

\item{package}{A string with the name of the R package to analyse the data, among the following ones: "rjags"}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the stochastic nodes to output from the analysis. By default all stochastic nodes are included.}

\item{inits}{aa}

\item{n.adapt}{aa}

\item{n.burnin}{aa}

\item{n.iter}{aa}

\item{thin}{aa}

\item{seed}{A positive integer specifying the random seed to use for analysing the data.}

\item{path}{A string specifying the path to the directory where the data sets are saved. By default \code{path = NULL } the data sets are not saved but are returned as an nlists object.}

\item{exists}{A flag specifying whether the directory should already exist. If \code{exists = NA} it doesn't matter. If the directory already exists it is overwritten if \code{exists = TRUE} or \code{exists = NA} otherwise an error is thrown.}

\item{silent}{A flag specifying whether to suppress warnings.}

\item{...}{not sure if will use}
}
\value{
A flag.
}
\description{
Analyse data for a simulation study. If path is supplied, saves a hidden file with the information on the analysis.
}
\examples{
dat <- sims::sims_simulate("a ~ dnorm(0,1)")
simanalyse_analyse_bayesian(dat, FUN=function(data, arg) data[arg], arg="a")
}
