% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-summarise.R
\name{simanalyse_summarise}
\alias{simanalyse_summarise}
\title{Summarises the results of simulation studies.}
\usage{
simanalyse_summarise(results.nlists, measures, estimator, monitor)
}
\arguments{
\item{results.nlists}{An nlists of results}

\item{measures}{A vector of strings indicating which Monte Carlo measures to calculate. Strings may include bias, rb (relative 
bias), br (bias ratio), var (variance), se (standard error), mse (mean square error), rmse (root mean square 
error), rrmse (relative root mean square error), cp90, cp95 and cp99 (coverage probability of 90, 95 and 99 percent quantile-based CrIs)}

\item{estimator}{A function, typically mean or median, for the Bayes estimator to use.}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the stochastic nodes in code to include in the summary. By default all stochastic nodes are included.}
}
\value{
A flag.
}
\description{
Summarises the results of simulation studies. 
The results are compared with the true parameter values by calculating Monte Carlo measures 
such as bias, mean square error and coverage probability.
R code can be used to customize the summary measures and to coerce results of analyses to
nlist when neccessary.
}
