% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-analyse-custom.R
\name{sma_analyse_custom}
\alias{sma_analyse_custom}
\title{Analyse data for a simulation study}
\usage{
sma_analyse_custom(datalist, FUN, path, seed, exists, silent, ...)
}
\arguments{
\item{datalist}{A list of nlist objects containing the data. Alternatively, path can be used to specify to read the data from files.}

\item{FUN}{An R function to analyse the data. It should contain an argument called data which takes an nlist of the data.}

\item{path}{A string specifying the path to the directory where the data sets are saved. By default \code{path = NULL } the data sets are not saved but are returned as an nlists object.}

\item{seed}{An integer specifying the random seed to use for analysing the data.}

\item{exists}{A flag specifying whether the directory should already exist. If \code{exists = NA} it doesn't matter. If the directory already exists it is overwritten if \code{exists = TRUE} or \code{exists = NA} otherwise an error is thrown.}

\item{silent}{A flag specifying whether to suppress warnings.}

\item{...}{Arguments to be passed to \code{FUN}.}
}
\value{
A flag.
}
\description{
Analyse data for a simulation study. If path is supplied, saves a hidden file with the information on the analysis.
}
\examples{
dat <- sims::sims_simulate("a ~ dnorm(0,1)")
sma_analyse_custom(dat, FUN=function(data, arg) data[arg], arg="a")
}
