% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-assess.R
\name{sma_assess}
\alias{sma_assess}
\title{Calculates posterior predictive checks (Bayesian p-values)}
\usage{
sma_assess(object, expr, data, code, constants = nlist(),
  nsamples = 100L, statistic = "FT")
}
\arguments{
\item{object}{An mcmcr object of results}

\item{expr}{An expression to calculate the expectations, e.g. "X = mu" if the expectation of X is mu}

\item{data}{A dataset}

\item{code}{Code to simulate data}

\item{constants}{Constants to pass with the code}

\item{nsamples}{Number of samples to take from the posterior distribution}

\item{statistic}{A string, one of "FT" (Freeman-Tukey), "LR" (Likelihood Ratio), or "chi2" (Chi-square)}
}
\value{
A flag.
}
\description{
Calculates posterior predictive checks (Bayesian p-values)
}
\examples{
code="for(i in 1:10){
a[i] ~ dpois(mu)}"
sims <- sims::sims_simulate(code, parameters=nlist(mu=5), nsims=1)
res <- sma_analyse_bayesian(sims, 
                           code, 
                           "mu ~ dunif(0,10)", 
                           monitor = "mu",
                           mode=sma_set_mode("debug", n.save=300, max.iter=300))
sma_assess(res, "a = mu", sims, code)
}
