% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-derive.R
\name{sma_derive}
\alias{sma_derive}
\title{Apply R code to each element of an nlist object. This can be used to derive new variables or create some missing data in the simulated data 
or in the results of the analyses.}
\usage{
sma_derive(object, code, monitor = ".*")
}
\arguments{
\item{object}{A nlists object}

\item{code}{A string of R code to modify the results.}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the stochastic nodes in code to include in the data. By default all stochastic nodes are included.}
}
\value{
A flag.
}
\description{
Derive new variables from the results of analyses
}
\examples{
set.seed(10L)
code <- "for(i in 1:10){x[i] ~ dnorm(mu,1)}"
dat <- sims::sims_simulate(code, parameters = nlist(mu=0), nsims=2)
sma_derive(dat, "y=x+1")
}
