% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-derive.R
\name{sma_derive_files}
\alias{sma_derive_files}
\title{Apply R code to derive new variables.}
\usage{
sma_derive_files(
  code,
  monitor = ".*",
  values = list(),
  path = ".",
  folder = "analysis0000001",
  progress = FALSE,
  options = furrr::furrr_options()
)
}
\arguments{
\item{code}{A string of R code to derive posterior samples for new parameters. E.g. "var = sigma^2".}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the variables in \code{object} and/or \code{code} to monitor. By default all variables are included.}

\item{values}{A named list of additional R objects to evaluate in the R expression.}

\item{path}{A string. Analyses results are read from that path on disk.}

\item{folder}{A string for the name of the folder that contains the analysis.}

\item{progress}{A flag specifying whether to print a progress bar.}

\item{options}{The future specific options to use with the workers.}
}
\value{
An object of the same class as \code{object}
}
\description{
Apply R code to derive new variables.
}
