% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_read_files.R
\name{sma_read_files}
\alias{sma_read_files}
\title{Read files}
\usage{
sma_read_files(fun, path = ".", folder = "analysis0000001", id = 1)
}
\arguments{
\item{fun}{The function from which to read the ouput file. Options are sma_analyse_files, sma_derive_files and sma_evaluate_files.}

\item{path}{A string. The object is read using this path.}

\item{folder}{A string for the name of the folder within the path that contains the analysis.}

\item{id}{An integer for the index of the object to extract.}
}
\value{
The object read from files.
}
\description{
Read output files
}
\examples{
params <- nlist(mu=0)
sims::sims_simulate("a ~ dnorm(mu, 1)", 
                    parameters = params, 
                    nsims = 5,
                    save = TRUE,
                    path=tempdir(),
                    exist = NA)
sma_analyse_files(code = "a ~ dnorm(mu, 1)
                          mu ~ dunif(-3,3)",
                  mode = sma_set_mode("quick"),
                  monitor = "mu",
                  path = tempdir())
sma_evaluate_files(parameters = params, 
                   path = tempdir())
sma_read_files(sma_evaluate_files, path = tempdir())
}
