% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_set_mode.R
\name{sma_set_mode}
\alias{sma_set_mode}
\title{Set mode for data analysis}
\usage{
sma_set_mode(mode = "reset")
}
\arguments{
\item{mode}{A string of the analysis mode.}
}
\description{
Set mode for data analysis
}
\details{
The possible modes are as follows:
\describe{
  \item{'debug'}{To rapidly identify problems with a model definition.}
  \item{'report'}{To produce results for a report.}
  \item{'paper'}{To produce results for a peer-reviewed paper.}
  \item{'reset'}{To reset all the options to NULL so that they are the default values for each function call.}
}

In each case the mode is a unique combination of the following package options
\describe{
  \item{sma.n.chains}{A count of the number of chains.}
  \item{sma.max.save}{A count of the number of simulations to save per chain.}
  \item{sma.max.iter}{A count specifying the maximum number of iterations.}
  \item{sma.rhat}{A number specifying the rhat threshold.}
  \item{sma.esr}{A number specifying the minimum effective sampling rate.}
  \item{sma.max.time}{The maximum total time to spend on analysis and reanalysis.}
}
}
\examples{
\dontrun{
sma_set_mode("debug")
}
}
