% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgtemplate-package.R
\docType{package}
\name{simanalyse-package}
\alias{simanalyse}
\alias{simanalyse-package}
\title{simanalyse: A Package to Facilitate Model Comparisons and Simulation Studies}
\description{
A Package to Facilitate Model Comparisons and Simulation Studies.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/audrey-b/simanalyse}
  \item Report bugs at \url{https://github.com/audrey-b/simanalyse/issues}
}

}
\author{
\strong{Maintainer}: Audrey Beliveau \email{audrey.beliveau@uwaterloo.ca}

Authors:
\itemize{
  \item Joe Thorley \email{joe@poissonconsulting.ca} (0000-0002-7683-4592)
}

}
\keyword{internal}
