% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-analyse-bayesian.R
\name{simanalyse_analyse_bayesian}
\alias{simanalyse_analyse_bayesian}
\title{Analyse data for a simulation study}
\usage{
simanalyse_analyse_bayesian(datalist, code, code_add = NULL,
  code_values = "", package = "rjags", monitor, inits = list(),
  n.adapt, n.burnin, n.iter, thin = 1, seed = sims::rcount())
}
\arguments{
\item{datalist}{A list of nlist objects containing the data. Alternatively, path can be used to specify to read the data from files.}

\item{code}{A string of code to analyse the data. JAGS code must not be in a data or model block.}

\item{code_add}{A string of code to add at the end of \code{code} before analysing the data. This is useful for specifying priors seperately from the likelihood in JAGS.}

\item{code_values}{A character vector to pass to sprintf when applied to the appended code from \code{code} and \code{code_add}}

\item{package}{A string with the name of the R package to analyse the data, among the following ones: "rjags"}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the stochastic nodes to output from the analysis. By default all stochastic nodes are included.}

\item{inits}{aa}

\item{n.adapt}{aa}

\item{n.burnin}{aa}

\item{n.iter}{aa}

\item{thin}{aa}

\item{seed}{A positive integer specifying the random seed to use for analysing the data.}
}
\value{
A flag.
}
\description{
Analyse data for a simulation study. If path is supplied, saves a hidden file with the information on the analysis.
}
\examples{
dat <- sims::sims_simulate("a ~ dnorm(0,1)")
}
