% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-analyse-bayesian.R
\name{simanalyse_analyse_bayesian}
\alias{simanalyse_analyse_bayesian}
\title{Analyse data for a simulation study}
\usage{
simanalyse_analyse_bayesian(datalist, code, code_add = "",
  code_values = "", package = "rjags", monitor, inits = list(),
  n.adapt, n.burnin, n.iter, thin = 1, n.chains = 3, seed = rcount())
}
\arguments{
\item{datalist}{A list of nlist objects containing the data. Alternatively, path can be used to specify to read the data from files.}

\item{code}{A string of code to analyse the data. JAGS code must not be in a data or model block.}

\item{code_add}{A string of code to add at the end of \code{code} before analysing the data. This is useful for specifying priors seperately from the likelihood in JAGS.}

\item{code_values}{A character vector to pass to sprintf when applied to the appended code from \code{code} and \code{code_add}. This is useful for varying choices of distributions or conducting sensitivity to prior analyses.}

\item{package}{A string with the name of the R package to analyse the data, among the following ones: "rjags"}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the stochastic nodes to output from the analysis. By default all stochastic nodes are included.}

\item{inits}{Initial values for the mcmc chains}

\item{n.adapt}{An integer specifying the number of adaptations for each analysis}

\item{n.burnin}{An integer specifying the number of burn-in iterations for each analysis (following the adaptation phase)}

\item{n.iter}{An integer specifying the number of iterations for each analysis (following the burn-in phase)}

\item{thin}{A numeric scalar of at least 1 specifying the thinning factor. Default is 1.}

\item{n.chains}{An integer greater than zero specifying the number of mcmc chains to run}

\item{seed}{A positive integer specifying the random seed to use for analysing the data.}
}
\value{
A flag.
}
\description{
Analyse data for a simulation study. If path is supplied, saves a hidden file with the information on the analysis.
}
\examples{
 set.seed(10L)
 dat <- sims::sims_simulate("a ~ dnorm(mu,1)", parameters = nlist(mu=0), nsims=2)
 result <- simanalyse_analyse_bayesian(datalist=dat,
 code = "a ~ dnorm(mu,1)
 mu ~ dunif(-3,3)",
 n.adapt = 101,
 n.burnin = 0,
 n.iter = 101,
 monitor = "mu")
}
