% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simanalyse-derive.R
\name{sma_derive}
\alias{sma_derive}
\title{Apply R code to each element of an nlist object. This can be used to derive new variables or create some missing data in the simulated data 
or in the results of the analyses.}
\usage{
sma_derive(object, code, monitor = ".*")
}
\arguments{
\item{object}{A single nlists object or a list of nlists objects.}

\item{code}{A string of R code to modify the nlists in \code{object}.}

\item{monitor}{A character vector (or regular expression if a string) specifying the names of the variables in \code{object} and/or \code{code} to monitor. By default all variables are included.}
}
\value{
A flag.
}
\description{
Derive new variables from the results of analyses
}
\examples{
set.seed(10L)
code <- "for(i in 1:10){x[i] ~ dnorm(0,1/variance)}"
dat <- sims::sims_simulate(code, parameters = nlist(variance=4), nsims=2)
res <- sma_analyse_bayesian(dat, code, code.add = "variance ~ dunif(0,10)", 
n.adapt=100, n.burnin=0, n.iter=3, monitor="variance")
sma_derive(res, "sd=sqrt(variance)")
}
