% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_set_mode.R
\name{sma_set_mode}
\alias{sma_set_mode}
\title{Set mode for data analysis}
\usage{
sma_set_mode(mode = "report", n.chains, n.adapt, max.save, max.iter,
  r.hat, esr, max.time, units, batch)
}
\arguments{
\item{mode}{A string of the analysis mode. One of "debug", "report" or "paper". The mode determines the values of the parameters below unless they are specified.}

\item{n.chains}{A count for number of chains.}

\item{n.adapt}{A count for the number of adaptations.}

\item{max.save}{A count for the number of maximum samples to save.}

\item{max.iter}{A count for the maximum number of iterations in total per chain.}

\item{r.hat}{A number specifying the rhat threshold.}

\item{esr}{A number specifying the minimum effective sampling rate.}

\item{max.time}{A number specifying the maximum time to spend on analysis.}

\item{units}{A character string specifying the units of time for \code{max.time}. See \code{difftime}.}

\item{batch}{A count specifying the number of iterations per batch.
The possible modes are as follows:
\describe{
  \item{'debug'}{To rapidly identify problems with a model definition.}
  \item{'report'}{To produce results for a report.}
  \item{'paper'}{To produce results for a peer-reviewed paper.}
}}
}
\description{
Set mode for data analysis
}
\examples{
\dontrun{
sma_set_mode("debug")
}
}
