% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sma_set_mode.R
\name{sma_set_mode}
\alias{sma_set_mode}
\title{Set mode for data analysis}
\usage{
sma_set_mode(mode = "report", n.chains, n.adapt, n.save, max.iter, r.hat,
  r.hat.nodes, esr, esr.nodes, max.time, units)
}
\arguments{
\item{mode}{A string of the analysis mode. One of "quick", "report" or "paper". The mode determines the values of the parameters below unless they are specified.}

\item{n.chains}{A count for number of chains.}

\item{n.adapt}{A count for the number of adaptations.}

\item{n.save}{A count for the number of (potentially thinned) samples to save.}

\item{max.iter}{A count for the maximum number of iterations in total per chain.}

\item{r.hat}{A number specifying the rhat threshold.}

\item{r.hat.nodes}{A character vector of nodes to be used to assess r.hat}

\item{esr}{A number specifying the minimum effective sampling rate.}

\item{esr.nodes}{A character vector of nodes to be used to assess esr}

\item{max.time}{A number specifying the maximum time to spend on analysis.}

\item{units}{A character string specifying the units of time for \code{max.time}. See \code{difftime}.
The possible modes are as follows:
\describe{
  \item{'quick'}{To rapidly identify problems with a model definition.}
  \item{'report'}{To produce results for a report.}
  \item{'paper'}{To produce results for a peer-reviewed paper.}
}}
}
\description{
Set mode for data analysis. The algorithm starts with an adaptation phase of \code{n.adapt} iterations.
Then \code{n.chains} are ran for \code{n.save} iterations without thinning. The algorithm stops if convergence
was achieved according to \code{r.hat} and the minimum effective sampling rate was achieved or \code{max.time}
was reached or is projected to be reached in the next round of iterations. If the algorithm is not stopped,
the iterations so far are discarded as burnin. The number of iterations is doubled from the previous round
and thinning is multiplied by 2. The stop criteria is applied again and so on.
}
\examples{
\dontrun{
sma_set_mode("quick")
}
}
